% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duck_latin.R
\name{duck_latin}
\alias{duck_latin}
\title{Duck Latin Function}
\usage{
duck_latin(.x)
}
\arguments{
\item{.x}{The word or list of words to be converted}
}
\value{
The converted word or list of words
}
\description{
Converts a word or list of words from english to duck latin
}
\examples{
word_list <- c("red", "orange", "yellow")
duck_latin(word_list)
}
