% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duck_latin.R
\name{duck_latin}
\alias{duck_latin}
\title{Duck Latin Function}
\usage{
duck_latin(.x)
}
\arguments{
\item{.x}{The word or list of words to be converted}
}
\value{
The converted word or list of words
}
\description{
Converts a word or list of words from english to a created langauge called duck latin. 
Duck latin translates from english by first moving the consonant cluster at the beginning of 
the word to the end. If the word begins with a vowel sound, no change is made. After, the sound 
"ak" is appended to the end of the word.
}
\examples{
word_list <- c("red", "orange", "yellow")
duck_latin(word_list)
}
